Initial commit
I made this project in the past 3 days, but I made it closed to the public because it is just for me and my group of friends.

I'm deciding to just copy paste it and enhance it as time passes because it is fun and I'd like to showcase it.